+++ test/support/test_feedback.ex
@@ -4,5 +4,7 @@ defmodule AshFeedback.Test.TestFeedback do
     domain: AshFeedback.Test.Feedback,
     repo: AshFeedback.Test.Repo,
     assignee_resource: AshFeedback.Test.User,
-    pubsub: :ash_feedback_test_pubsub
+    pubsub: :ash_feedback_test_pubsub,
+    audio_blob_resource: AshFeedback.Test.StorageBlob,
+    audio_attachment_resource: AshFeedback.Test.StorageAttachment
 end